backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.18
openpyxl==3.1.5
PyPDF2==3.0.1
python-docx==1.1.2
PyJWT>=2.8.0
bcrypt>=4.0.0
